fix(arm): make a failed gain write terminal for that servo
The D->I->P write order was justified as a safety property -- every prefix
at least as stable as the state before it -- but the loop logged and
continued, so the reachable states were all eight subsets rather than the
four prefixes. {I,P} without D leaves servo 3 at 48/32/4, the exact
combination the ordering exists to prevent: measured to oscillate, in
EEPROM, on a gravity-loaded joint, with no retry to clear it
(applyServoGains returns nothing, so initializeServosWithRetry never
re-runs for a gain failure).

01d57ea widened the window by attempting all three writes even when the
reads fail, so the break is now the only thing bounding the failure set.
Per servo, not per call.

The order had no test: reversing the slice left the whole suite green.
Adds a wire-level recorder, since final register values are
order-independent. Gains fields become byte so an out-of-range retune is a
compile error rather than silent truncation to less damping, and the
invariant test pins the report's caveat that servo 2 oscillated at I=4 in
2 of 3 repeat passes.

+++ internal/servo/gains.go
@@ -1,10 +1,13 @@
 package servo
 
-// Gains is one servo's position-loop PID register triple. Each value is a single byte.
+// Gains is one servo's position-loop PID register triple. byte, not int, so a retune out of
+// register range is a compile error rather than a silent truncation on the wire -- a D of
+// 300 would reach the servo as 44, i.e. LESS damping than intended, which is the direction
+// that breaks applyServoGains' write-order argument.
 type Gains struct {
-	P int
-	D int
-	I int
+	P byte
+	D byte
+	I byte
 }
 
 // DefaultArmGains is the tuned per-joint gain table for arm servos 1-5, against a stock
